Work around osx/plist error with Ruby 2.0.
Apparently, TextMate's plist extension is incompatible with Ruby 2.0. This
causes the error: "An object in the argument tree could not be converted (ArgumentError)"
when a command that displays a dialog is executed. That plist extension does not
seem actively developed (the latest commits are from about 4 years ago)...

Starting with OS X 10.9 (aka Mavericks), Ruby 2.0 is the default system Ruby.
Fortunately, Apple has kept Ruby 1.8 around (at least for now).

So, to work around TM's plist bug, we force all commands to use Ruby 1.8. This is
not an ideal solution, because it will break as soon as Apple drops the support
for 1.8 or changes the path to the frameworks. But, by that time, hopefully TM
will have a working plist extension again.